requirements.txt
Babel==1.3
beautifulsoup4==4.4.1
Django==1.8.4
django-appconf==1.0.1
django-compressor==1.5
django-extra-views==0.6.4
django-haystack==2.3.1
django-modelcluster==0.6.2
django-oscar==1.1.1
django-tables2==1.0.4
django-taggit==0.17.1
django-treebeard==3.0
django-widget-tweaks==1.3
djangorestframework==3.2.4
factory-boy==2.4.1
html5lib==0.999
mock==1.0.1
phonenumbers==6.3.0
Pillow==2.7.0
purl==1.1
pycountry==1.15
pytz==2015.6
six==1.9.0
sorl-thumbnail==12.2
Unidecode==0.4.18
wagtail==1.1
Willow==0.2.1
